refactor "fun" package to "misc"
