+++ src/main/java/me/mcofficer/james/James.java
@@ -8,7 +8,9 @@ import me.mcofficer.james.audio.Audio;
 import me.mcofficer.james.commands.*;
 import me.mcofficer.james.commands.audio.*;
 import me.mcofficer.james.commands.creatortools.*;
-import me.mcofficer.james.commands.fun.*;
+import me.mcofficer.james.commands.misc.Birb;
+import me.mcofficer.james.commands.misc.Cat;
+import me.mcofficer.james.commands.misc.Dog;
 import me.mcofficer.james.commands.info.*;
 import me.mcofficer.james.commands.lookup.*;
 import me.mcofficer.james.commands.moderation.*;
